Create ha40_laundryRoom
uses 2 accelerometers to monitor a washer and dryer. the accelerometers operate through the I2C bus, and each is hard-coded to have the exact same address. as a result, they are wired through a mux which allows them to be read separately. for more details on the setup, check out: https://joshuarobotprojects.wordpress.com/