Create kookiePicFood.py
this is a python script that runs via terminal command. it accepts an input from an arduino program which uses an IR sensor and photoresistor to detect the cat and ensure proper lighting for the picture. if the proper conditions are met, it takes the picture and tweets it.

for more details, check out my blog: https://joshuarobotprojects.wordpress.com/